Правки функций, модуляции
