+++ js/main.js
@@ -1,3 +1,7 @@
-import './random-function.js';
 import './utils.js';
-import './data.js';
+import './mock-data.js';
+import {getMockAds} from './mock-data.js';
+
+getMockAds(10);
+// const data = getMockAds(10);
+// console.log(data);
